Remove OpenSearch multimodal component
Delete the legacy flows/components/opensearch_multimodal.py implementation (large multimodal OpenSearch vector store component). The changes consolidate/remove the old multimodal implementation and update flows/components/opensearch_multimodel.py accordingly (refactor/rename work). Update callers/references as needed since the previous multimodal file has been removed.
